Rename worker modules to services
